Moved to docker-compose for multiple applications running.
Added a sample database API looking to format it to specific use case in LER 2.0.

Dropped cargo-chef as in built cargo vendor is sufficient in caching in docker builds.

Applications are moving steadily along. LER-2.0 Server is on hold until certificate is in place.

Will call LER 2.0 for test certificate for unblocking devlopment.